Added setCanvasDimensions in the Canvas component.
The canvas component now gets its current size and sets it to local state then calls
and updates the redux store.
The Tail component outputs the canvas dimensions with the intention that later
you can input new dimensions here in order to change the size of the element.
Also a resize listener is placed on the canvas element.