Redo on the store 'canvasClickCoordinates'.
This method is now removed and replaced with two methods ( 'canvasMouseDownCoordinates'
and 'canvasMouseUpCoordinates' ) each handles its respective mouse event coordinates and
sets it on a property on the store.
Right now the coordinate system for the app is { x: int, y: int } - which is an object with
x and y properties, but it may change to [ x, y ] - which would be an array with 0 index being
x and 1 index being y.