Added a form to the ShapeMenu component.
The ShapeMenu component will house the basic shapes 'circle, square and triangle'
and the form controls various attributes for the shape. This form is wired to the
store through the shapeActions.js action creator and put into the general store.
It will be used from the Canvas component to draw the selected shape with the corresponding
attributes.